Testing optimisation for packets -Downcasting position, rotation and other information to int8 or int 16

+++ Multiplayer_CA2/Multiplayer_CA2/MultiplayerGameState.hpp
@@ -42,7 +42,7 @@ class MultiplayerGameState : public State
 		TextureHolder&				mTextureHolder;
 
 		std::map<int, PlayerPtr>	mPlayers;
-		std::vector<sf::Int32>		mLocalPlayerIdentifiers;
+		std::vector<sf::Int8>		mLocalPlayerIdentifiers;
 		sf::TcpSocket				mSocket;
 		bool						mConnected;
 		std::unique_ptr<GameServer> mGameServer;
